+++ src/hooks/useDuplicateSimulation.ts
@@ -42,6 +42,7 @@ export function useDuplicateSimulation() {
           modelId: currentSimulation.modelId,
           layerIdByMaterialId: currentSimulation.layerIdByMaterialId,
           solverSettings: currentSimulation.solverSettings,
+          taskType: currentSimulation.taskType,
         }).unwrap();
 
         await updateSimulation({
@@ -55,6 +56,7 @@ export function useDuplicateSimulation() {
             solverSettings: currentSimulation.solverSettings,
             sources: currentSimulation.sources,
             receivers: currentSimulation.receivers,
+            taskType: currentSimulation.taskType,
           },
         }).unwrap();